Extract GA4 and UA specific logic form Google API
