+++ lib/plausible/google/api.ex
@@ -1,9 +1,13 @@
-defmodule Plausible.Google.Api do
-  alias Plausible.Google.{GA4, UA, HTTP}
-  use Timex
-  require Logger
+defmodule Plausible.Google.API do
+  @moduledoc """
+  API to Google services.
+  """
 
-  @type google_analytics_view() :: {view_name :: String.t(), view_id :: String.t()}
+  use Timex
+
+  alias Plausible.Google.HTTP
+
+  require Logger
 
   @search_console_scope URI.encode_www_form(
                           "email https://www.googleapis.com/auth/webmasters.readonly"
@@ -23,6 +27,10 @@ defmodule Plausible.Google.Api do
 
     "https://accounts.google.com/o/oauth2/v2/auth?client_id=#{client_id()}&redirect_uri=#{redirect_uri()}&prompt=consent&response_type=code&access_type=offline&scope=#{@import_scope}&state=" <>
       Jason.encode!([site_id, redirect_to, legacy, ga4])
+  end
+
+  def fetch_access_token!(code) do
+    HTTP.fetch_access_token!(code)
   end
 
   def fetch_verified_properties(auth) do
@@ -56,228 +64,7 @@ defmodule Plausible.Google.Api do
     end
   end
 
-  @spec list_views(access_token :: String.t()) ::
-          {:ok, %{(hostname :: String.t()) => [google_analytics_view()]}} | {:error, term()}
-  @doc """
-  Lists Google Analytics views grouped by hostname.
-  """
-  def list_views(access_token) do
-    case UA.HTTP.list_views_for_user(access_token) do
-      {:ok, %{"items" => views}} ->
-        views = Enum.group_by(views, &view_hostname/1, &view_names/1)
-        {:ok, views}
-
-      error ->
-        error
-    end
-  end
-
-  def list_properties(access_token) do
-    case GA4.HTTP.list_accounts_for_user(access_token) do
-      {:ok, %{"accountSummaries" => accounts}} ->
-        accounts =
-          accounts
-          |> Enum.filter(& &1["propertySummaries"])
-          |> Enum.map(fn account ->
-            {"#{account["displayName"]} (#{account["account"]})",
-             Enum.map(account["propertySummaries"], fn property ->
-               {"#{property["displayName"]} (#{property["property"]})", property["property"]}
-             end)}
-          end)
-
-        {:ok, accounts}
-
-      error ->
-        error
-    end
-  end
-
-  defp view_hostname(view) do
-    case view do
-      %{"websiteUrl" => url} when is_binary(url) -> url |> URI.parse() |> Map.get(:host)
-      _any -> "Others"
-    end
-  end
-
-  defp view_names(%{"name" => name, "id" => id}) do
-    {"#{id} - #{name}", id}
-  end
-
-  @spec get_view(access_token :: String.t(), lookup_id :: String.t()) ::
-          {:ok, google_analytics_view()} | {:ok, nil} | {:error, term()}
-  @doc """
-  Returns a single Google Analytics view if the user has access to it.
-  """
-  def get_view(access_token, lookup_id) do
-    case list_views(access_token) do
-      {:ok, views} ->
-        view =
-          views
-          |> Map.values()
-          |> List.flatten()
-          |> Enum.find(fn {_name, id} -> id == lookup_id end)
-
-        {:ok, view}
-
-      {:error, cause} ->
-        {:error, cause}
-    end
-  end
-
-  def get_property(access_token, lookup_property) do
-    case list_properties(access_token) do
-      {:ok, properties} ->
-        property =
-          properties
-          |> Enum.map(&elem(&1, 1))
-          |> List.flatten()
-          |> Enum.find(fn {_name, property} -> property == lookup_property end)
-
-        {:ok, property}
-
-      {:error, cause} ->
-        {:error, cause}
-    end
-  end
-
-  @type import_auth :: {
-          access_token :: String.t(),
-          refresh_token :: String.t(),
-          expires_at :: String.t()
-        }
-
-  @per_page 7_500
-  @backoff_factor :timer.seconds(10)
-  @max_attempts 5
-  @spec import_analytics(Date.Range.t(), String.t(), import_auth(), (String.t(), [map()] -> :ok)) ::
-          :ok | {:error, term()}
-  @doc """
-  Imports stats from a Google Analytics UA view to a Plausible site.
-
-  This function fetches Google Analytics reports in batches of #{@per_page} per
-  request. The batches are then passed to persist callback.
-
-  Requests to Google Analytics can fail, and are retried at most
-  #{@max_attempts} times with an exponential backoff. Returns `:ok` when
-  importing has finished or `{:error, term()}` when a request to GA failed too
-  many times.
-
-  Useful links:
-
-  - [Feature documentation](https://plausible.io/docs/google-analytics-import)
-  - [GA API reference](https://developers.google.com/analytics/devguides/reporting/core/v4/rest/v4/reports/batchGet#ReportRequest)
-  - [GA Dimensions reference](https://ga-dev-tools.web.app/dimensions-metrics-explorer)
-
-  """
-  def import_analytics(date_range, view_id, auth, persist_fn) do
-    with {:ok, access_token} <- maybe_refresh_token(auth) do
-      do_import_analytics(date_range, view_id, access_token, persist_fn)
-    end
-  end
-
-  defp do_import_analytics(date_range, view_id, access_token, persist_fn) do
-    Enum.reduce_while(UA.ReportRequest.full_report(), :ok, fn report_request, :ok ->
-      report_request = %UA.ReportRequest{
-        report_request
-        | date_range: date_range,
-          view_id: view_id,
-          access_token: access_token,
-          page_token: nil,
-          page_size: @per_page
-      }
-
-      case fetch_and_persist(report_request, persist_fn: persist_fn) do
-        :ok -> {:cont, :ok}
-        {:error, _} = error -> {:halt, error}
-      end
-    end)
-  end
-
-  @spec fetch_and_persist(UA.ReportRequest.t(), Keyword.t()) ::
-          :ok | {:error, term()}
-  def fetch_and_persist(%UA.ReportRequest{} = report_request, opts \\ []) do
-    persist_fn = Keyword.fetch!(opts, :persist_fn)
-    attempt = Keyword.get(opts, :attempt, 1)
-    sleep_time = Keyword.get(opts, :sleep_time, @backoff_factor)
-
-    case UA.HTTP.get_report(report_request) do
-      {:ok, {rows, next_page_token}} ->
-        :ok = persist_fn.(report_request.dataset, rows)
-
-        if next_page_token do
-          fetch_and_persist(
-            %UA.ReportRequest{report_request | page_token: next_page_token},
-            opts
-          )
-        else
-          :ok
-        end
-
-      {:error, cause} ->
-        if attempt >= @max_attempts do
-          {:error, cause}
-        else
-          Process.sleep(attempt * sleep_time)
-          fetch_and_persist(report_request, Keyword.merge(opts, attempt: attempt + 1))
-        end
-    end
-  end
-
-  def import_ga4_analytics(date_range, property, auth, persist_fn) do
-    with {:ok, access_token} <- maybe_refresh_token(auth) do
-      do_import_ga4_analytics(date_range, property, access_token, persist_fn)
-    end
-  end
-
-  defp do_import_ga4_analytics(date_range, property, access_token, persist_fn) do
-    Enum.reduce_while(GA4.ReportRequest.full_report(), :ok, fn report_request, :ok ->
-      report_request = %GA4.ReportRequest{
-        report_request
-        | date_range: date_range,
-          property: property,
-          access_token: access_token,
-          offset: 0,
-          limit: @per_page
-      }
-
-      case fetch_and_persist_ga4(report_request, persist_fn: persist_fn) do
-        :ok -> {:cont, :ok}
-        {:error, _} = error -> {:halt, error}
-      end
-    end)
-  end
-
-  @spec fetch_and_persist_ga4(GA4.ReportRequest.t(), Keyword.t()) ::
-          :ok | {:error, term()}
-  def fetch_and_persist_ga4(%GA4.ReportRequest{} = report_request, opts \\ []) do
-    persist_fn = Keyword.fetch!(opts, :persist_fn)
-    attempt = Keyword.get(opts, :attempt, 1)
-    sleep_time = Keyword.get(opts, :sleep_time, @backoff_factor)
-
-    case GA4.HTTP.get_report(report_request) do
-      {:ok, {rows, row_count}} ->
-        :ok = persist_fn.(report_request.dataset, rows)
-
-        if report_request.offset + @per_page < row_count do
-          fetch_and_persist(
-            %GA4.ReportRequest{report_request | offset: report_request.offset + @per_page},
-            opts
-          )
-        else
-          :ok
-        end
-
-      {:error, cause} ->
-        if attempt >= @max_attempts do
-          {:error, cause}
-        else
-          Process.sleep(attempt * sleep_time)
-          fetch_and_persist_ga4(report_request, Keyword.merge(opts, attempt: attempt + 1))
-        end
-    end
-  end
-
-  defp maybe_refresh_token(%Plausible.Site.GoogleAuth{} = auth) do
+  def maybe_refresh_token(%Plausible.Site.GoogleAuth{} = auth) do
     with true <- needs_to_refresh_token?(auth.expires),
          {:ok, {new_access_token, expires_at}} <- do_refresh_token(auth.refresh_token),
          changeset <-
@@ -293,7 +80,7 @@ defmodule Plausible.Google.Api do
     end
   end
 
-  defp maybe_refresh_token({access_token, refresh_token, expires_at}) do
+  def maybe_refresh_token({access_token, refresh_token, expires_at}) do
     with true <- needs_to_refresh_token?(expires_at),
          {:ok, {new_access_token, _expires_at}} <- do_refresh_token(refresh_token) do
       {:ok, new_access_token}
