+++ lib/plausible_web/templates/site/settings_search_console.html.heex
@@ -82,7 +82,7 @@
     <% else %>
       <PlausibleWeb.Components.Google.button
         id="search-console-connect"
-        to={Plausible.Google.Api.search_console_authorize_url(@site.id, "search-console")}
+        to={Plausible.Google.API.search_console_authorize_url(@site.id, "search-console")}
       />
       <div class="text-gray-700 dark:text-gray-300 mt-8">
         NB: You also need to set up your site on
