+++ test/plausible/google/api_test.exs
@@ -1,8 +1,8 @@
-defmodule Plausible.Google.ApiTest do
+defmodule Plausible.Google.APITest do
   use Plausible.DataCase, async: true
   use Plausible.Test.Support.HTTPMocker
 
-  alias Plausible.Google.Api
+  alias Plausible.Google
   alias Plausible.Imported.UniversalAnalytics
 
   import ExUnit.CaptureLog
@@ -44,7 +44,7 @@ defmodule Plausible.Google.ApiTest do
       Plausible.Imported.Buffer.insert_many(buffer, table, records)
     end
 
-    assert :ok == Plausible.Google.Api.import_analytics(date_range, view_id, auth, persist_fn)
+    assert :ok == Google.UA.API.import_analytics(date_range, view_id, auth, persist_fn)
 
     Plausible.Imported.Buffer.flush(buffer)
     Plausible.Imported.Buffer.stop(buffer)
@@ -86,7 +86,7 @@ defmodule Plausible.Google.ApiTest do
       Plausible.Imported.Buffer.insert_many(buffer, table, records)
     end
 
-    assert :ok == Plausible.Google.Api.import_analytics(range, "123551", auth, persist_fn)
+    assert :ok == Google.UA.API.import_analytics(range, "123551", auth, persist_fn)
 
     Plausible.Imported.Buffer.flush(buffer)
     Plausible.Imported.Buffer.stop(buffer)
@@ -127,7 +127,7 @@ defmodule Plausible.Google.ApiTest do
                 hideValueRanges: true,
                 metrics: [%{expression: "ga:users"}, %{expression: "ga:exits"}],
                 orderBys: [%{fieldName: "ga:date", sortOrder: "DESCENDING"}],
-                pageSize: 10000,
+                pageSize: 10_000,
                 pageToken: nil,
                 viewId: "123"
               }
@@ -139,7 +139,7 @@ defmodule Plausible.Google.ApiTest do
       )
 
       assert :ok =
-               Api.fetch_and_persist(request,
+               Google.UA.API.fetch_and_persist(request,
                  sleep_time: 0,
                  persist_fn: fn dataset, row ->
                    assert dataset == "imported_exit_pages"
@@ -178,7 +178,7 @@ defmodule Plausible.Google.ApiTest do
       }
 
       assert {:error, :request_failed} =
-               Api.fetch_and_persist(request,
+               Google.UA.API.fetch_and_persist(request,
                  sleep_time: 0,
                  persist_fn: fn _dataset, _rows -> :ok end
                )
@@ -209,7 +209,7 @@ defmodule Plausible.Google.ApiTest do
       }
 
       assert :ok ==
-               Api.fetch_and_persist(request,
+               Google.UA.API.fetch_and_persist(request,
                  sleep_time: 0,
                  persist_fn: fn dataset, rows ->
                    assert dataset == "imported_exit_pages"
@@ -253,7 +253,7 @@ defmodule Plausible.Google.ApiTest do
 
       query = %Plausible.Stats.Query{date_range: Date.range(~D[2022-01-01], ~D[2022-01-05])}
 
-      assert {:error, "google_auth_error"} = Plausible.Google.Api.fetch_stats(site, query, 5)
+      assert {:error, "google_auth_error"} = Google.API.fetch_stats(site, query, 5)
     end
 
     test "returns whatever error code google returns on API client error", %{site: site} do
@@ -270,7 +270,7 @@ defmodule Plausible.Google.ApiTest do
 
       query = %Plausible.Stats.Query{date_range: Date.range(~D[2022-01-01], ~D[2022-01-05])}
 
-      assert {:error, "some_error"} = Plausible.Google.Api.fetch_stats(site, query, 5)
+      assert {:error, "some_error"} = Google.API.fetch_stats(site, query, 5)
     end
 
     test "returns generic HTTP error and logs it", %{site: site} do
@@ -290,7 +290,7 @@ defmodule Plausible.Google.ApiTest do
       log =
         capture_log(fn ->
           assert {:error, "failed_to_list_stats"} =
-                   Plausible.Google.Api.fetch_stats(site, query, 5)
+                   Google.API.fetch_stats(site, query, 5)
         end)
 
       assert log =~ "Google Analytics: failed to list stats: %Finch.Error{reason: :some_reason}"
@@ -314,7 +314,7 @@ defmodule Plausible.Google.ApiTest do
               [
                 %{name: ["keyword1", "keyword2"], visitors: 25},
                 %{name: ["keyword3", "keyword4"], visitors: 15}
-              ]} = Plausible.Google.Api.fetch_stats(site, query, 5)
+              ]} = Google.API.fetch_stats(site, query, 5)
     end
 
     test "returns next page when page argument is set", %{user: user, site: site} do
@@ -336,7 +336,7 @@ defmodule Plausible.Google.ApiTest do
               [
                 %{name: ["keyword1", "keyword2"], visitors: 25},
                 %{name: ["keyword3", "keyword4"], visitors: 15}
-              ]} = Plausible.Google.Api.fetch_stats(site, query, 5)
+              ]} = Google.API.fetch_stats(site, query, 5)
     end
 
     test "defaults first page when page argument is not set", %{user: user, site: site} do
@@ -355,7 +355,7 @@ defmodule Plausible.Google.ApiTest do
               [
                 %{name: ["keyword1", "keyword2"], visitors: 25},
                 %{name: ["keyword3", "keyword4"], visitors: 15}
-              ]} = Plausible.Google.Api.fetch_stats(site, query, 5)
+              ]} = Google.API.fetch_stats(site, query, 5)
     end
 
     test "returns error when token refresh fails", %{user: user, site: site} do
@@ -372,7 +372,7 @@ defmodule Plausible.Google.ApiTest do
 
       query = %Plausible.Stats.Query{date_range: Date.range(~D[2022-01-01], ~D[2022-01-05])}
 
-      assert {:error, "invalid_grant"} = Plausible.Google.Api.fetch_stats(site, query, 5)
+      assert {:error, "invalid_grant"} = Google.API.fetch_stats(site, query, 5)
     end
   end
 
@@ -393,7 +393,7 @@ defmodule Plausible.Google.ApiTest do
             %{
               "one.test" => [{"57238190 - one.test", "57238190"}],
               "two.test" => [{"54460083 - two.test", "54460083"}]
-            }} == Plausible.Google.Api.list_views("access_token")
+            }} == Google.UA.API.list_views("access_token")
   end
 
   test "list_views/1 returns authentication_failed when request fails with HTTP 403" do
@@ -405,7 +405,7 @@ defmodule Plausible.Google.ApiTest do
       end
     )
 
-    assert {:error, :authentication_failed} == Plausible.Google.Api.list_views("access_token")
+    assert {:error, :authentication_failed} == Google.UA.API.list_views("access_token")
   end
 
   test "list_views/1 returns authentication_failed when request fails with HTTP 401" do
@@ -417,7 +417,7 @@ defmodule Plausible.Google.ApiTest do
       end
     )
 
-    assert {:error, :authentication_failed} == Plausible.Google.Api.list_views("access_token")
+    assert {:error, :authentication_failed} == Google.UA.API.list_views("access_token")
   end
 
   test "list_views/1 returns error when request fails with HTTP 500" do
@@ -429,6 +429,6 @@ defmodule Plausible.Google.ApiTest do
       end
     )
 
-    assert {:error, :unknown} == Plausible.Google.Api.list_views("access_token")
+    assert {:error, :unknown} == Google.UA.API.list_views("access_token")
   end
 end
